% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rintrinio.R
\name{gather_financial_statement_company_compare}
\alias{gather_financial_statement_company_compare}
\title{Given the tickers, statement, year and period returns all the information from the Intrinio API fundamental reported financials
  for that time and those tickers as a dataframe}
\usage{
gather_financial_statement_company_compare(
  api_key,
  ticker,
  statement,
  year,
  period
)
}
\arguments{
\item{api_key}{character API key (sandbox or production) from Intrinio}

\item{ticker}{character vector of the ticker symbols you would like to study}

\item{statement}{character the statement that you want to study
options: 'income_statement', 'cash_flow_statement', 'balance_sheet_statement'}

\item{year}{character the year you want the information from}

\item{period}{character the period you want the information from}
}
\value{
a dataframe that contains information about the given statement for the given tickers at the given time
}
\description{
Given the tickers, statement, year and period returns all the information from the Intrinio API fundamental reported financials
  for that time and those tickers as a dataframe
}
\examples{
gather_financial_statement_company_compare('OjhlMjhjNTBmY2IyMWJiMWE0MTExYjQwNWZmZTVkZWM1', c('AAPL', 'CSCO'), 'income_statement', '2019', 'Q1')
}
