% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rintrinio.R
\name{gather_financial_statement_time_series}
\alias{gather_financial_statement_time_series}
\title{Given the ticker, statement, year(s) and period(s) return the complete available financial information from the Intrinio API stock data}
\usage{
gather_financial_statement_time_series(
  api_key,
  ticker,
  statement,
  year,
  period
)
}
\arguments{
\item{api_key}{character (sandbox or production) from Intrinio}

\item{ticker}{character the ticker symbol you would like to get information for}

\item{statement}{character that represents the financial statement that you want to study
options: 'income_statement', 'cash_flow_statement', 'balance_sheet_statement'}

\item{year}{character vector of the year(s) you want the information for}

\item{period}{character vector of the period(s) you want the information for}
}
\value{
dataframe containing information about the given statement for the given ticker at the given times
}
\description{
Given the ticker, statement, year(s) and period(s) return the complete available financial information from the Intrinio API stock data
}
\examples{
gather_financial_statement_time_series('OjhlMjhjNTBmY2IyMWJiMWE0MTExYjQwNWZmZTVkZWM1',
'CVX', 'income_statement', c('2017','2018'), c('Q1','Q3'))
}
