% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rintrinio.R
\name{gather_stock_returns}
\alias{gather_stock_returns}
\title{Given the tickers, buy-in date, sell-out date, returns the historical prices and profit/loss}
\usage{
gather_stock_returns(api_key, ticker, buy_date, sell_date)
}
\arguments{
\item{api_key}{character API key (sandbox or production) from Intrinio}

\item{ticker}{character ticker symbols or a vector of ticker symbols}

\item{buy_date}{character the buy-in date in the format of "%Y-%m-%d", e.g. "2019-12-31"
If the input date is not a trading day, it will be automatically changed to the next nearest trading day.}

\item{sell_date}{character the sell-out date in the format of "%Y-%m-%d", e.g. "2019-12-31"
If the input date is not a trading day, it will be automatically changed to the last nearest trading day.}
}
\value{
a dataframe that contains the companies, historical prices and corresponding
profit/loss
}
\description{
Given the tickers, buy-in date, sell-out date, returns the historical prices and profit/loss
}
\examples{
gather_stock_returns('OjhlMjhjNTBmMWJiMWE0MTExYjQwNWZmZTVkZWM1', c('AAPL', 'CSCO'),
"2017-12-31", "2019-03-01")
}
