% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rintrinio.R
\name{gather_stock_time_series}
\alias{gather_stock_time_series}
\title{Given the ticker, start date, and end date, return from the Intrinio API stock data
for that time frame in either a dictionary or a pandas dataframe format.}
\usage{
gather_stock_time_series(
  api_key,
  ticker,
  start_date = "",
  end_date = "",
  allow_max_rows = FALSE
)
}
\arguments{
\item{api_key}{character API key (sandbox or production) from Intrinio}

\item{ticker}{character the ticker symbol you would like to get stock data for}

\item{start_date}{character optional the earliest date in the format of "%Y-%m-%d",
e.g. "2019-12-31" to get data for}

\item{end_date}{character optional the most recent date in the format of "%Y-%m-%d",
e.g. "2019-12-31" to get data for}

\item{allow_max_rows}{boolean optional if False, then only 100 rows will show in the
output, otherwise up to 10000 rows will show (based on dates)}
}
\value{
a dataframe that contains stock data for the specified timeframe
}
\description{
Given the ticker, start date, and end date, return from the Intrinio API stock data
for that time frame in either a dictionary or a pandas dataframe format.
}
\examples{
gather_stock_time_series('OjhlMjhjNTBmY2IyMWJiMWE0MTExYjQwNWZmZTVkZWM1', 'AAPL',
"2017-12-31", "2019-03-01")
}
